Partially implemented the Progress Screen, changed double relative to [0, 100] int

+++ src/main/java/com/feed_the_beast/ftbquests/quest/tasks/QuestTask.java
@@ -59,7 +59,7 @@ public abstract class QuestTask extends QuestObject
 	}
 
 	@Override
-	public final double getRelativeProgress(ITeamData data)
+	public final int getRelativeProgress(ITeamData data)
 	{
 		return data.getQuestTaskData(this).getRelativeProgress();
 	}
